Added source to ACR workflow instructions
This PR focuses around ironing out the flow of running `spk project init` or
`spk project init -m`, creating pipelines via the generated
`azure-pipelines.yaml` files, and ensuring the pipeline can create appropriately
named/tagged docker images in ACR.

In order to allow for flexibility in how SRE's choose to deploy the pipelines,
this PR does not automate the actual deployment of the pipeline. Instead, `info`
level instructions have been added specifying how to run the
`spk service create-pipeline` to deploy the `azure-pipelines.yaml` generated; If
the the user has configured the spk config correctly, the outputted command is
copy/paste ready to deploy the generated pipelines yaml:

If ran on a mono-repo:
`spk service create-pipeline --project-dir packages my-service` will be logged.

If ran on a standard repo: `spk service create-pipeline my-service` will be
logged.

The `azure-pipelines.yaml` expects 4 environment variables to be present in
order to login to `az` cli and build and push an image to ACR:

- `ACR_NAME` - name of the target ACR
- `SP_APP_ID` - service principal ID
- `SP_PASS` - service principal password
- `SP_TENANT` - service principal tenant

These requirements are also `info` logged to the end user upon
`azure-pipelines.yaml` generation.

Previously `spk service create-pipeline` only worked for mono-repositories and
the code has been refactored such that the existence of the `--packages-dir`
options implies whether or not the repository is a mono-repository or a standard
one.

closes https://github.com/microsoft/bedrock/issues/646