Added command `spk ingress-route create`
Added the `spk ingress-route create` command which allows users to create
Traefik `IngressRoute` manifests for the Traefik 2 IngressRoute CRD.

The command is setup such that a target `<service>` and `<port>` are required;
these will be used together to target a service in the Kubernetes cluster
matching `<service>:<port>`. If a `--ring` is provided, the target service
becomes `<service>-<ring>:<port>`.

The YAML output can either be outputted to a file with the `--output-file`
option, the option allows the user to specify any path-like directory on the
host; in the case of relative directories, it is resolved from the `cwd` of spk.
If a user leaves the `--output-file` blank or provides and empty string, the
YAML is printed to `stdout` with no additional log messages; this is useful if
one needs to pipe the output to in a pipeline.

closes https://github.com/microsoft/bedrock/issues/619